Fix unsigned/signed comparison

+++ src/builtin.cc
@@ -31,7 +31,7 @@ std::vector<double> Filter(FilterFn fn, std::vector<double> input) {
 Collection Map(MapFn fn, Collection input) {
   // TODO: thread
   double* output = new double[input.length];
-  for (int i = 0; i < input.length; ++i)
+  for (size_t i = 0; i < input.length; ++i)
     output[i] = fn(input.values[i]);
 
   return {output, input.length};
